+++ build.sbt
@@ -21,6 +21,8 @@ val jUnitRuntime = Build.jUnitRuntime
 val jUnitTestOutputsJS = Build.jUnitTestOutputsJS
 val jUnitTestOutputsJVM = Build.jUnitTestOutputsJVM
 val jUnitPlugin = Build.jUnitPlugin
+val jsDependenciesCore = Build.jsDependenciesCore
+val jsDependenciesPlugin = Build.jsDependenciesPlugin
 val examples = Build.examples
 val helloworld = Build.helloworld
 val reversi = Build.reversi
