+++ jsdependencies-core/src/main/scala/org/scalajs/jsdependencies/core/DependencyResolver.scala
@@ -1,4 +1,4 @@
-package org.scalajs.core.tools.jsdep
+package org.scalajs.jsdependencies.core
 
 import scala.collection.mutable
