+++ jsdependencies-sbt-plugin/src/main/scala/org/scalajs/jsdependencies/sbtplugin/AbstractJSDeps.scala
@@ -1,10 +1,10 @@
-package org.scalajs.sbtplugin
+package org.scalajs.jsdependencies.sbtplugin
 
 import sbt._
 
 import StringUtilities.nonEmpty
 
-import org.scalajs.core.tools.jsdep.JSDependency
+import org.scalajs.jsdependencies.core.JSDependency
 
 /** Something JavaScript related a project may depend on. Either a JavaScript
  *  module/library, or the DOM at runtime. */
