+++ sbt-plugin-test/project/build.sbt
@@ -1,2 +1,5 @@
 addSbtPlugin("org.scala-js" % "sbt-scalajs" %
   org.scalajs.core.ir.ScalaJSVersions.current)
+
+addSbtPlugin("org.scala-js" % "sbt-scalajs-jsdependencies" %
+  org.scalajs.core.ir.ScalaJSVersions.current)
